% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bbm_to_abm}
\alias{bbm_to_abm}
\alias{bbm_to_angle}
\title{Convert bbm (matrix of bounding boxes) to elements of slope-intercept matrix or vector of angles}
\usage{
bbm_to_abm(m)

bbm_to_angle(m, m2 = NULL)
}
\arguments{
\item{m, m2}{matrix, where columns represent x1, y1, x2, y2 of bbx}
}
\value{
matrix of slopes and intercepts

vector of angles
}
\description{
Convert bbm (matrix of bounding boxes) to elements of slope-intercept matrix or vector of angles
}
\examples{
m <-bbx_to_bbm(c("0 0 5 5", "5 5 10 10"))
bbm_to_abm(m)

m <-bbx_to_bbm(c("0 0 5 5", "5 5 10 10"))
bbm_to_angle(m)

}
