% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbx-package.R
\docType{package}
\name{bbx-package}
\alias{bbx}
\alias{bbx-package}
\title{bbx: Frame your world}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Utility functions for operating with bounding boxes and geometries for magick images.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmi3kno/bbx}
  \item Report bugs at \url{https://github.com/dmi3kno/bbx/issues}
}

}
\author{
\strong{Maintainer}: Dmytro Perepolkin \email{dperepolkin@gmail.com} (\href{https://orcid.org/0000-0001-8558-6183}{ORCID})

}
\keyword{internal}
