% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{bbx_aggregate}
\alias{bbx_aggregate}
\alias{bbx_union}
\alias{bbx_intersect}
\title{Functions for aggregating bbx objects
These functions can perform union and intersection operations on bbx objects}
\usage{
bbx_aggregate(bbx, bbx2 = NULL, fx1, fy1, fx2, fy2)

bbx_union(bbx, bbx2 = NULL)

bbx_intersect(bbx, bbx2 = NULL)
}
\arguments{
\item{bbx}{character vector of bounding boxes to perform operation on}

\item{bbx2}{optional character vector of bounding boxes to element-wise aggregation with `bbx`.
If specified, needs to be length 1 or equal in length to `bbx`.}

\item{fx1, fy1, fx2, fy2}{functions to use for aggregating x1, y1, x2, y2, respectively. Defaults to `min` for x1,y1 and `max` for x2,y2}
}
\value{
bbx or missing value, if result is invalid bounding box
}
\description{
Functions for aggregating bbx objects
These functions can perform union and intersection operations on bbx objects
}
\examples{
bbx_union(c("5 1 7 3", "2 4 6 8"))
bbx_union(c("5 1 7 3", "2 4 6 8"), c("1 1 1 1"))
bbx_intersect(c("5 1 7 3", "2 4 6 8")) # should return NA
bbx_intersect("5 1 7 3", "2 2 6 8")
}
