% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{bbx_to_geometry}
\alias{bbx_to_geometry}
\alias{geometry_to_bbx}
\title{Convert bbx to magick-geometry}
\usage{
bbx_to_geometry(bbx)

geometry_to_bbx(g)
}
\arguments{
\item{bbx}{character vector of bounding boxes to be converted to geometries}

\item{g}{character vector of geometries to be converted to bbxes}
}
\value{
character vector of magick-compliant geometry string representing bounding box area in the format  "width x height +x_off +y_off" (https://www.imagemagick.org/Magick++/Geometry.html)
}
\description{
Convert bbx to magick-geometry
}
\examples{
# area from c(0,0) to c(100,200) counting from top left
bbx_to_geometry("0 0 100 200")
bbx_to_geometry(c("0 0 100 200", "100,100,200, 200"))
geometry_to_bbx(c("100x200+12+14", "100x200", "+12+14", "x200+12+14"))
}
\seealso{
[magick::geometry] for converting integers (four individual columns) to geometry
}
