% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbx.R
\name{bbx_area}
\alias{bbx_area}
\alias{bbx_h}
\alias{bbx_w}
\alias{bbx_x}
\alias{bbx_y}
\title{Functions for accessing bbx properties
These functions can calculate various metrix of bbx}
\usage{
bbx_area(bbx)

bbx_h(bbx)

bbx_w(bbx)

bbx_x(bbx)

bbx_y(bbx)
}
\arguments{
\item{bbx}{character vector of bounding boxes to pad}
}
\value{
a vector of values
}
\description{
Functions for accessing bbx properties
These functions can calculate various metrix of bbx
}
\details{
Functions `bbx_h` and `bbx_w` return height and width, respectively.
Functions `bbx_x` and `bbx_y` return `x` and `y` offset.
}
\examples{
bbx_area("100 100 200 200")
bbx_h("100 100 200 200")
bbx_w("100 100 200 200")
}
