% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bbx_to_rlst}
\alias{bbx_to_rlst}
\alias{bbx_to_clst}
\alias{rlst_to_bbx}
\alias{clst_to_bbx}
\title{Convert bbx to row/column list and back}
\usage{
bbx_to_rlst(bbx)

bbx_to_clst(bbx)

rlst_to_bbx(l)

clst_to_bbx(l)
}
\arguments{
\item{bbx}{character vector of bbxes}

\item{l}{list of numeric vectors, representing bbxes or bbx coordinates (for `bbx_to_rlst()` and `bbx_to_clst()`, respectively)}
}
\value{
list numeric vectors representing bbxes

list numeric vectors representing bbx coordinates

character vector of bbxes

character vector of bbxes
}
\description{
Convert bbx to row/column list and back
}
\examples{
l <- c("0 0 5 5", "5,5,10,10")
bbx_to_rlst(l)
bbx_to_clst(l)
l <- list(c(0,0,5,5), c(5,0,10,5), c(0,5,5,10), c(5,5,10,10))
rlst_to_bbx(l)
clst_to_bbx(l)
}
