% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{bbx_area}
\alias{bbx_area}
\title{Functions for calculating with bbx
These functions can calculate various metrix of bbx}
\usage{
bbx_area(bbx)
}
\arguments{
\item{bbx}{character vector of bounding boxes to pad}
}
\value{
a vector of validated bbxes
}
\description{
Functions for calculating with bbx
These functions can calculate various metrix of bbx
}
\examples{
bbx_area("100 100 200 200")

}
