% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bbx_to_bbm}
\alias{bbx_to_bbm}
\alias{bbm_to_bbx}
\alias{coords_to_bbx}
\title{Convert bbx to matrix and matrix/vectors to bbx}
\usage{
bbx_to_bbm(bbx)

bbm_to_bbx(m)

coords_to_bbx(x1, y1, x2, y2)
}
\arguments{
\item{bbx}{character vector of bbxes (top-left coordinates). Expects four integers separated by comma or space}

\item{m}{matrix, where columns represent x1, y1, x2, y2 of bbx}

\item{x1, y1, x2, y2}{vectors representing x1, y1, x2, y2 of bbx}
}
\value{
character vector of bbxes
}
\description{
Convert bbx to matrix and matrix/vectors to bbx
}
\examples{
bbx_to_bbm(c("0 0 5 5", "5 5 10 10"))
bbm_to_bbx(matrix(c(0,0,5,5,5,5,10,10), nrow=2, byrow=TRUE))

coords_to_bbx(c(0,5,0,5),c(0,0,5,5),
               c(5,10,5,10), c(5,5,10,10))

}
