% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbx.R
\name{bbx_pad_width}
\alias{bbx_pad_width}
\alias{bbx_pad_height}
\title{Functions for padding bbx
These functions can "pad" (increase size of) bbx}
\usage{
bbx_pad_width(bbx, n = 1, word = NULL, side = "both")

bbx_pad_height(bbx, n = 1, word = NULL, side = "both")
}
\arguments{
\item{bbx}{character vector of bounding boxes to pad}

\item{n}{integer number of pixels to add. If a `word` is provided, `n` is interpreted as number of characters.}

\item{word}{optional character vector of words contained in bbxes}

\item{side}{"left", "right" (for `bbx_pad_width()`), "up", "down" (for `bbx_pad_height()`) or "both" which side to pad}
}
\value{
a vector of validated bbxes
}
\description{
Functions for padding bbx
These functions can "pad" (increase size of) bbx
}
\examples{
bbx_pad_width("5 5 10 20", word="There")
bbx_pad_width("5 5 10 20", 1)
bbx_pad_height("5 5 10 20", word="There/nbe/ndragons")
bbx_pad_height("5 5 10 20", 1)
}
