% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbx.R
\name{bbx_intersects}
\alias{bbx_intersects}
\title{Predicate functions for matching bbxes
These functions can check whether intersection operation on bbx objects returns non-NA result}
\usage{
bbx_intersects(bbx, bbx2 = NULL)
}
\arguments{
\item{bbx}{character vector of bounding boxes to perform operation on}

\item{bbx2}{optional character vector of bounding boxes to element-wise aggregation with `bbx`.
If specified, needs to be length 1 or equal in length to `bbx`.}
}
\value{
logical value of whether or not the pair of bbxes intersect
}
\description{
Predicate functions for matching bbxes
These functions can check whether intersection operation on bbx objects returns non-NA result
}
\examples{
bbx_intersects(c("5 1 7 3", "2 4 6 8")) # should return FALSE
bbx_intersects("5 1 7 3", "2 2 6 8") # should return TRUE
}
