% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbx.R
\name{bbx_slice_x}
\alias{bbx_slice_x}
\alias{bbx_slice_y}
\alias{bbx_slice_xy}
\title{Functions for horizontal and vertical slicing of bbx columns
These functions are not vectorized and should be used to compute}
\usage{
bbx_slice_x(bbx, x)

bbx_slice_y(bbx, y)

bbx_slice_xy(bbx, x, y)
}
\arguments{
\item{bbx}{string or character vector of bbx'es}

\item{x, y}{coordinates to sclice the bouding box at}
}
\value{
list of bbx'es appliccable for particular split
}
\description{
Functions for horizontal and vertical slicing of bbx columns
These functions are not vectorized and should be used to compute
}
\examples{
bbx_slice_x("0 0 100 200", 80)
bbx_slice_x(c("0 0 100 200", "100 100 200 200"), c(80, 150))

bbx_slice_y("0 0 100 200", 120)
bbx_slice_y(c("0 0 100 200", "100 100 200 200"), c(120,150))
bbx_slice_xy("0 0 100 200", 50, 100)
bbx_slice_xy(c("0 0 100 200", "100,100, 200, 200"), c(50, 150), c(100, 150))
}
