% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{bbx_translate}
\alias{bbx_translate}
\title{Functions for transforming certain coordinates of bbx
These functions can modify a vector of bbx}
\usage{
bbx_translate(bbx, x = NULL, y = NULL)
}
\arguments{
\item{bbx}{character vector of bounding boxes to update}

\item{x}{scalar or numeric vector to update bbx with}

\item{y}{scalar or numeric vector to update bbx with}
}
\value{
a vector of updated and validated bbxes
}
\description{
Functions for transforming certain coordinates of bbx
These functions can modify a vector of bbx
}
\examples{
bbx_translate(c("100 100 200 200", "300 400 500 600"), x=111, y=222)

}
