% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{bbx_is_valid}
\alias{bbx_is_valid}
\alias{bbx_validate}
\title{Functions for validating bbx
These functions can check whether specified bbx is valid, i.e. x1 <= x2 and y1 <= y2}
\usage{
bbx_is_valid(bbx)

bbx_validate(bbx)
}
\arguments{
\item{bbx}{character vector bounding boxes to validate}
}
\value{
a vector of logical values

original vector with NA for invalid bbxes
}
\description{
Functions for validating bbx
These functions can check whether specified bbx is valid, i.e. x1 <= x2 and y1 <= y2
}
\examples{
bbx_is_valid("0 0 100 200")
bbx_validate(c("5,4,6,3", "1,1,5,6"))
}
