% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{bbx_intersects}
\alias{bbx_intersects}
\alias{bbx_ovelaps}
\title{Predicate functions for matching bbxes}
\usage{
bbx_intersects(bbx, bbx2)

bbx_ovelaps(bbx, bbx2, direction = "x")
}
\arguments{
\item{bbx}{character vector of bounding boxes to perform operation on}

\item{bbx2}{optional character vector of bounding boxes to element-wise aggregation with `bbx`.
If specified, needs to be length 1 or equal in length to `bbx`.}

\item{direction}{by default x}
}
\value{
logical value
}
\description{
Predicate functions for matching bbxes
}
\details{
These functions can check intersection(non-empty intersection geometry) or
 horizontal/vertical overlap of segments.
}
\examples{
bbx_intersects("5 1 7 3", "2 2 6 8") # should return TRUE
bbx_ovelaps("100 100 200 200", "120, 220, 180, 240") # should be TRUE

}
